Ruff format all files

+++ nicos_mlz/dns/devices/detector.py
@@ -32,25 +32,33 @@ from nicos.devices.generic.detector import Detector
 from nicos.devices.polarized.flipper import OFF, ON, BaseFlipper
 from nicos.devices.tango import PyTangoDevice
 
-P_TIME = 't'
-P_TIME_SF = 'tsf'
-P_TIME_NSF = 'tnsf'
-P_MON = 'mon1'
-P_MON_SF = 'mon1sf'
-P_MON_NSF = 'mon1nsf'
+P_TIME = "t"
+P_TIME_SF = "tsf"
+P_TIME_NSF = "tnsf"
+P_MON = "mon1"
+P_MON_SF = "mon1sf"
+P_MON_NSF = "mon1nsf"
 
 
 class TofChannel(TOFChannel):
     """Basic Tango Device for TofDetector."""
 
     parameters = {
-        'readchannels': Param('Tuple of (start, end) channel numbers will be '
-                              'returned by a read', type=tupleof(int, int),
-                              default=(0, 0), settable=True, mandatory=True),
-        'readtimechan': Param('Tuple of (start, end) integrated time channels '
-                              'will be returned by a read',
-                              type=tupleof(int, int),
-                              default=(0, 0), settable=True, mandatory=True),
+        "readchannels": Param(
+            "Tuple of (start, end) channel numbers will be " "returned by a read",
+            type=tupleof(int, int),
+            default=(0, 0),
+            settable=True,
+            mandatory=True,
+        ),
+        "readtimechan": Param(
+            "Tuple of (start, end) integrated time channels "
+            "will be returned by a read",
+            type=tupleof(int, int),
+            default=(0, 0),
+            settable=True,
+            mandatory=True,
+        ),
     }
 
     def doInit(self, mode):
@@ -58,7 +66,7 @@ class TofChannel(TOFChannel):
         if mode != SIMULATION:
             self._dev.set_timeout_millis(10000)
             shape = tuple(self._dev.detectorSize)[::-1]
-        self.arraydesc = ArrayDesc(self.name, shape=shape, dtype='<u4')
+        self.arraydesc = ArrayDesc(self.name, shape=shape, dtype="<u4")
 
     def doPrepare(self):
         self._dev.Clear()
@@ -81,22 +89,24 @@ class TofChannel(TOFChannel):
 
     def valueInfo(self):
         start, end = self.readchannels
-        return tuple(Value("chan-%d" % i, unit="cts", errors="sqrt",
-                           type="counter", fmtstr="%d")
-                     for i in range(start, end + 1))
+        return tuple(
+            Value("chan-%d" % i, unit="cts", errors="sqrt", type="counter", fmtstr="%d")
+            for i in range(start, end + 1)
+        )
 
     def doReadArray(self, quality):
         self.arraydesc = ArrayDesc(
-            self.name, shape=tuple(self._dev.detectorSize)[::-1], dtype='<u4')
+            self.name, shape=tuple(self._dev.detectorSize)[::-1], dtype="<u4"
+        )
         start, end = self.readchannels
         # get current data array from detector
         array = numpy.asarray(self._dev.value, numpy.uint32)
         array = array.reshape(self.arraydesc.shape)
         if self.timechannels > 1:
             startT, endT = self.readtimechan
-            res = array[startT:endT+1].sum(axis=0)[start:end+1]
+            res = array[startT : endT + 1].sum(axis=0)[start : end + 1]
         else:
-            res = array[0, start:end+1]
+            res = array[0, start : end + 1]
         self.readresult = res.tolist()
         return array
 
@@ -105,13 +115,16 @@ class DNSDetector(Detector):
     """Detector supporting different presets for spin flipper on or off."""
 
     attached_devices = {
-        'flipper': Attach('Spin flipper device which will be read out '
-                          'with respect to setting presets.', BaseFlipper),
+        "flipper": Attach(
+            "Spin flipper device which will be read out "
+            "with respect to setting presets.",
+            BaseFlipper,
+        ),
     }
 
     def _getWaiters(self):
         waiters = self._adevs.copy()
-        del waiters['flipper']
+        del waiters["flipper"]
         return waiters
 
     def doTime(self, preset):
@@ -137,9 +150,13 @@ class DNSDetector(Detector):
                 m = preset[P_MON_NSF]
             new_preset = {P_MON: m}
         elif P_MON_SF in preset or P_MON_NSF in preset:
-            self.log.warning('Incorrect preset setting. Specify either both '
-                             '%s and %s or only %s.',
-                             P_MON_SF, P_MON_NSF, P_MON)
+            self.log.warning(
+                "Incorrect preset setting. Specify either both "
+                "%s and %s or only %s.",
+                P_MON_SF,
+                P_MON_NSF,
+                P_MON,
+            )
             return
         elif P_TIME_SF in preset and P_TIME_NSF in preset:
             if self._attached_flipper.read() == ON:
@@ -148,9 +165,13 @@ class DNSDetector(Detector):
                 t = preset[P_TIME_NSF]
             new_preset = {P_TIME: t}
         elif P_TIME_SF in preset or P_TIME_NSF in preset:
-            self.log.warning('Incorrect preset setting. Specify either both '
-                             '%s and %s or only %s.',
-                             P_TIME_SF, P_TIME_NSF, P_TIME)
+            self.log.warning(
+                "Incorrect preset setting. Specify either both "
+                "%s and %s or only %s.",
+                P_TIME_SF,
+                P_TIME_NSF,
+                P_TIME,
+            )
             return
         elif P_MON in preset:
             new_preset = {P_MON: preset[P_MON]}
